refactor(ui): P3 pilot — admin components + website/ui retirement onto base-nova
Pattern-proof sweep establishing the canonical mapping table before fan-out:
- admin org-list/org-sheet/user-sheet/styled-sub-data-grid: Badge import →
  reui/badge (PRD A1 amendment: vendored base badge has only 6 chrome
  variants, registry badge carries the semantic set the app needs);
  SwitchWrapper → plain div (dropped export, context was internal-only);
  intent="plain"/size="sq-sm" → variant="ghost"/size="icon-sm"
- components/website/ui/{button,command,popover}.tsx DELETED (radix-based
  copies whose sole consumer was ui/multi-selector)
- multi-selector rewired onto vendored ui/button+popover+command:
  asChild → render, redundant manual toggle removed (Base UI triggers
  toggle natively), onEscapeKeyDown dropped, classNames verbatim

Pilot files tsc-clean; total burn-down 304 → 294.
